res enhanced the mixmode abilities a bit. res added blending support to SR.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@22680 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/video/render3d/shader/shaderplugins/softshader/softshader_fp.cpp
@@ -104,9 +104,9 @@ void csSoftShader_FP::SetupState (const csRenderMesh* mesh,
     65536.0f, 1.0f/32768.0f);
   int shift;
   if (f >= 1.0)
-    shift = csLog2 (f);
+    shift = csLog2 ((int)f);
   else
-    shift = -csLog2 (1.0f/f);
+    shift = -csLog2 ((int)(1.0f/f));
   shaderPlug->scanlineRenderer->SetShift (shift);
 }
